Fix "unresolved symbol: memset" on release builds.
This was being caused by recent growth of our PERFECT_HASH_TLS_CONTEXT
structure resulting in the following construct being replaced by memset
for release builds:

    PERFECT_HASH_TLS_CONTEXT LocalTlsContext = { 0 };

Solution: don't zero initialize the structure, instead, clear it on-demand
if needed in PerfectHashTlsGetOrSetContext().

As this change required an inline (i.e. non-Rtl) way to zero memory, we also
introduce a new ZeroMemoryInline routine and helper macros, and convert the
applicable areas in GraphReset() to use the new macros.