Implement initial "find best graph" behavior.
This differs from the normal "first graph wins" behavior in that
a "best" predicate can be specified, and the graph solving innards
will keep generating solutions and capturing the "best" one until
a configurable limit is hit.

This commit also introduces table create parameters, which I'd been
putting off for a while.  More work needs to be done to process the
command line parameters and convert them into the appropriate C
structs.