Implement support for specifying seed mask byte counts.
This is used to generate weighted random bytes for seed masks (which
would have been captured from a previous run).  Experiments show that
this will lead to higher solving probabilities in some situations
(especially when in find best graph mode and best coverage attempts is
greater than or equal to 2).