Implement the Multiply643ShiftR and Multiply644ShiftR hash functions.
These two functions have identical implementations, the only thing that differs
is the masks used for seed 3's byte 1 and 2.  643 uses a seed mask of 0x1f1f
(0x1f = 32, hence the '3' in 643), whereas 644 uses a seed mask of 0x3f3f (0x3f
= 64, hence the 4 in 644).

The routine differs from MultiplyShiftR in that it uses 64-bit intermediates for
the result of the multiplication.