Implement --AutoResizeWhenKeysToEdgesRatioExceeds=D.
Supplies a keys to edges ratio that, if exceeded, results in an auto
resize, i.e. the equivalent of --InitialNumberOfTableResizes=1.  Valid
values are above 0.0 and less than 1.0.  Typical values would be 0.8 to
0.9 depending on the hash function being used.

This will result in much faster solving rates for "nearly full" key sets
(i.e., when the number of keys is very close to a power of two, e.g.
 HydrogenWorld-31016.keys).