Implement a vertex collision database.
Capture the keys that result in vertex collisions, and, when sufficient
collisions have been encountered, switch into a solving mode where the
most frequently collided keys are hashed first.

This was an experiment to see if there's any runtime advantage in
focusing on a subset of keys during hashing.

It didn't seem to make much difference to the rate at which solving
attempts were being made for keys like HologramWorld-31016.keys.  And
there weren't nearly as many hits as I was expecting in the "vertex
collision from db/from LRU" lines I hacked into the console status
output (the number slowly crept up in tens/fifties, I would have liked
to have seen it account for say 50% of vertex collision avoidances).

So, commiting for posterity, but not planning on merging for now.