Implement MultiplyShiftRX.
This is identical to MultiplyShiftR, but uses a fixed final right-shift
amount that is automatically set to the value of 32 minus the number of
trailing zeros of the number of vertices (i.e. the power of 2 exponent
that yields the number of vertices -> 2 ** 15 = 32768; 32 - 15 = 17,
thus the seed 3 byte 1 and 2 values will be hard-coded to 17).

This results in faster final versions of the compiled hash routines, as
the two vertex AND operations can be omitted.

N.B. This was stumbled upon accidentally when reviewing the disassembly
     of a compiled table's index routine that was inexplicably faster
     than index routines of identical runs.  The faster table had a
     vertex count of 65536, and had randomly resulted in seed 3 byte 1
     and 2 values of 0x10.  MSVC figured out that the final AND by the
     hash mask, 0xffff, was unnecessary, due to the preceeding >> 0x10,
     and omitted the redundant instructions (in release mode).  Thus,
     the resulting assembly featured less instructions, and thus, ran
     a little faster (a few cycles) than the other compiled versions
     that *did* do the final AND masking.  (Also of note, neither gcc
     or clang did that optimization in release builds of the same table.)