Tweak MultiplyShiftRX() logic.
Shift right by SEED3_BYTE1 for both vertices; they have the exact same
value.

Additionally, remove the logic that reduces the seed count if a hash
function indicates it doesn't need final AND masking; this was a tiny
optimization that needs to be undone in preparation for the upcoming
Mulshrolate hash functions.